+++ src/Enums/ValidationRuleEnum.php
@@ -40,4 +40,14 @@ enum ValidationRuleEnum: string
      * The field under validation must be a valid Universally Unique Lexicographically Sortable Identifier (ULID).
      */
     public const ULID = 'ulid';
+
+    /**
+     * The field under validation must be included in the given list of values in:foo,bar...
+     */
+    public const IN = 'in:';
+
+    /**
+     * The field under validation must not be included in the given list of values not_in:foo,bar,bar...
+     */
+    public const NOT_IN = 'not_in:';
 }