+++ CHANGELOG.md
@@ -8,3 +8,5 @@ All notable changes to `shergela/validation-rule` will be documented in this fil
 - add new `endsWith()` method
 - add new `doesntStartWith()` method
 - add new `doesntEndWith()` method
+- add new `in()` method
+- add new `notIn()` method